+++ src/main/java/interface_adapter/presenter/TimerPresenter.java
@@ -2,10 +2,10 @@ package interface_adapter.presenter;
 
 import interface_adapter.viewmodel.SetTimerViewModel;
 import interface_adapter.viewmodel.RunningTimerViewModel;
-import use_case.SetTimerUseCase.SetTimerOutputBoundary;
-import use_case.SetTimerUseCase.SetTimerOutputData;
-import use_case.UpdateTimerUseCase.UpdateTimerOutputBoundary;
-import use_case.UpdateTimerUseCase.UpdateTimerOutputData;
+import use_case.TimerUseCases.SetTimerUseCase.SetTimerOutputBoundary;
+import use_case.TimerUseCases.SetTimerUseCase.SetTimerOutputData;
+import use_case.TimerUseCases.UpdateTimerUseCase.UpdateTimerOutputBoundary;
+import use_case.TimerUseCases.UpdateTimerUseCase.UpdateTimerOutputData;
 
 /**
  * Presenter for the timer, implementing the output boundaries for setting the timer.
