--- src/main/java/use_case/PauseTimerUseCase/PauseTimerInputBoundary.java
@@ -1,4 +1,0 @@
-package use_case.PauseTimerUseCase;
-
-public interface PauseTimerInputBoundary {
-}
