--- src/main/java/use_case/PauseTimerUseCase/PauseTimerInputData.java
@@ -1,4 +1,0 @@
-package use_case.PauseTimerUseCase;
-
-public class PauseTimerInputData {
-}
