--- src/main/java/use_case/PauseTimerUseCase/PauseTimerOutputBoundary.java
@@ -1,4 +1,0 @@
-package use_case.PauseTimerUseCase;
-
-public interface PauseTimerOutputBoundary {
-}
