+++ src/main/java/use_case/TimerUseCases/PauseTimerUseCase/PauseTimerInputData.java
@@ -1,0 +1,4 @@
+package use_case.TimerUseCases.PauseTimerUseCase;
+
+public class PauseTimerInputData {
+}
