+++ src/main/java/use_case/TimerUseCases/UpdateTimerUseCase/UpdateTimerInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.UpdateTimerUseCase;
+package use_case.TimerUseCases.UpdateTimerUseCase;
 
 public interface UpdateTimerInputBoundary {
     void execute(UpdateTimerInputData updateTimerInputData);
