+++ src/main/java/use_case/TimerUseCases/UpdateTimerUseCase/UpdateTimerOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.UpdateTimerUseCase;
+package use_case.TimerUseCases.UpdateTimerUseCase;
 
 public interface UpdateTimerOutputBoundary {
     void prepareSuccessView(UpdateTimerOutputData outputData);
